+++ debug-adapter/app/Phoityne/VSCode/TH/VariableJSON.hs
@@ -16,6 +16,7 @@ data Variable =
     nameVariable               :: String  -- The variable's name
   , typeVariable               :: String  -- he variable's type.
   , valueVariable              :: String  -- The variable's value. For structured objects this can be a multi line text, e.g. for a function the body of a function.
+  , evaluateNameVariable       :: Maybe String  -- Optional evaluatable name of this variable which can be passed to the 'EvaluateRequest' to fetch the variable's value.
   , variablesReferenceVariable :: Int     -- If variablesReference is > 0, the variable is structured and its children can be retrieved by passing variablesReference to the VariablesRequest.
   } deriving (Show, Read, Eq)